pyproject.toml
[tool.poetry]
name = "query-doc"
version = "0.1.0"
description = "This project is based on Documents Understanding library Donut, integrated with a Streamlit frontend."
authors = ["davidzuma <zumaquerodavid@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.2.1"
opencv-python = "^4.9.0.80"
transformers = "^4.38.1"
datasets = "^2.17.1"
sentencepiece = "^0.2.0"
pillow = "^10.2.0"
protobuf = "^4.25.3"
streamlit = "^1.31.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
